Split single file SSH into modules
